Remove flags acting on whole stack
