+++ shared/test/src/vyxal/YamlTests.scala
@@ -1,6 +1,5 @@
 package vyxal
 
-import scala.collection.mutable
 import scala.io.Source
 
 import org.scalatest.funspec.AnyFunSpec
